Move page helpers into utility file
